Use short_path when collecting paths to source files
Full path consist of an optional first part called `root`, and the second
part which is the `short_path`. Usually for non-generated files `path`
is equal to `short_path`, but this is not the case for the generated files.

Documentation states that the `short_path` should be used if the file is used
in the runfiles to handle both generated and standard files correctly.

Internal-tag: [#46586]
Signed-off-by: Robert Winkler <rwinkler@antmicro.com>