Improve python dependency handling in cocotb_test rule
The changes made in this commit include improvements that allow
the `py_library` targets to be properly passed to the `cocotb_test` rule via
the `deps` attribute. Now all transitive imports are properly added
to `PYTHONPATH`.

The `test_module` attribute, on the other hand,
is intended to contain only python source files.

Internal-tag: [#46586]
Signed-off-by: Robert Winkler <rwinkler@antmicro.com>