Rework the standard cell library support in `bazel_rules_hdl`.
 * Normalize temperature (`m40c` / `125c`) and voltage strings (`0v124`,
   `0v000`) found in corners.

 * Add object which records the various standard cell options
   (`StandardCellOptionsInfo`).

 * Add function which produces a list of standard cell libraries
   (`for_each_XXXX_cells` like `for_each_sky130_cells` or
   `for_each_asap7_cells`).

While it is expected that we will be using flows which use multi-VTs
together and multi-corner analysis, this type of functionality can make
it easier to separate out how various options effect the results of your
design. It is also great for determining a "performance envelope" that
your results will be contained within.

Signed-off-by: Tim 'mithro' Ansell <me@mith.ro>
Signed-off-by: Tim Ansell <tansell@google.com>