Remove noise from the PerformancePowerAreaProto package name.
Put it in hdl.ppa.PerformancePowerAreaProto.

A proto package name is meant to provide a namespace for the proto-buffer
and should not necessarily indicate specific filename path elements.

Previously, the package sounded like it is specifically tied to
bazel_rules_hdl; but it has nothing to do with bazel rules, just
a way to express data in the context of hdl.

Let's keep it simple and just call the package 'hdl.ppa'.