Refine proto buffer describing PPA results.
This standardizes units for power (watts), time (picoseconds),
and length units (um or um2) and breaks out more information that
might be available from the EDA tools.

Compared to the original proto, there are so many changes, that
this is a new protocol buffer, based on the old one.
So to not clash with the original, this uses a new name
PowerPerformanceAreaProto; the old proto needed a small
change in the name of a type (because we really want to use
the name 'Power' in the new one with a different definition),
but it is fully binary compatible to the one before.

TL;DR:
  * old: hdl.ppa.PerformancePowerAreaProto
  * new: hdl.ppa.PowerPerformanceAreaProto

Next step will be to adapt the bazel rules currently generating
the old proto to create the new one.

Once this new one is established, the old one can be
deprecated and removed quickly after last use.