Extract the correct power totals from the logfile.
There is another 'Total' in the logfile that would override
the values just found (and does not have multiple columns).

Make the regex a bit more specific to only latch on the line that
actually contains all the power values.

Also: 'Complex' combinational cells is now as 'Multi-input' combinational
cells in the log file. Match that as well.