Fix text protocol buffer: place `estimation_method` at right level.
Make textproto parseable.

According to [the schema](https://github.com/hdl/bazel_rules_hdl/blob/main/synthesis/power_performance_area.proto#L200),
the estimation_method is part of the 'Power' message, not a sub-element of the
PowerBreakdown message.

Also: estimation_method is a string, so it needs to be quoted.