verilog: adds support for include headers
Include headers should not appear in the verilator command
line, but should be available in the compilation sandbox.

It seems that verilator rules did not allow `hdrs` files into
the sandbox, so this change fixes the issue.

Furthermore, very often core authors use the include directive
to read an include file from the directory where the include
file is located:

```
`include "file.h"
```

This will not work in bazel, because the expected include path is:

```
`include "directory/where/file/resides/file.h"
```

This rule adds the `includes` parameter to `verilog_library` to allow
the user to specify a set of include directories, therefore allowing
verilation of included cores.

The `verilator_cc_library` rules is fixed to suport this. It is also
fixed to use the `hdrs` files as inputs to the verilation rule.